requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.2
pydantic-settings==2.1.0
pyyaml==6.0.1
bertopic==0.16.1
sentence-transformers==2.2.2
numpy==1.26.2
scikit-learn==1.3.2
streamlit==1.29.0
pandas==2.1.4
plotly==5.18.0
hdbscan>=0.8.29
umap-learn>=0.5.3
